Final_project/requirements.txt
gradio
langchain
openai
transformers
faiss-cpu
gemini-api
weaviate-client 
langchain 
tiktoken 
pypdf 
rapidocr-onnxruntime 
sentence-transformers 
streamlit
langchain-groq
langchain-community
langchain-huggingface
unstructured
langchain-google-genai
pdfminer.six
unstructured-pytesseract
tesseract-ocr
SpeechRecognition
moviepy
langchain-cohere
google-search-results